test(wallet): fixes transaction service tests stack overflow (#4243)
Description
---
- Fixes stack overflow in `discovery_async_return_test`
- Removes creation of custom runtime for each test

Motivation and Context
---
CI was reporting a stack overflow in `discovery_async_return_test`. 
All async tests should use `tokio::test`. The use of a std::sync::CondVar caused tests to fail when used in an async thread,
this was replaced with a tokio watch.

How Has This Been Tested?
---
Tests pass