feat(comms): add or_optional trait extension for RpcStatus (#4246)
Description
---
- feat(comms): add or_optional trait extension for RpcStatus
- fix(comms): fixup lifetimes which error in latest rust

Motivation and Context
---
When RPC returns NotFound it is common that a user wants to handle that error differently. This PR makes that more ergonomic by adding the `or_optional` function to the RPC result that returns an Option if the NotFound status is returned.  An example where NotFound this needs to be handled is in wallet utxo scanning. 

The lifetime changes in tor fix a compilation error that is likely from the recent upgrade to nightly rust. It is not clear why this error did not crop up sooner.

How Has This Been Tested?
---
New unit test