feat: add tari engine for flow and wasm functions (#4237)
A prototype and stub for a potential WASM and flow engine for DAN layer smart contracts.

Not fully functional, but wanted to get the code in. Most of the work is moving the StateDb and Instructions into the engine crate. The idea is that the `engine` crate can be referenced by template creators as a `dev_dependency` in their projects so that they can run tests without all the HotStuff consensus code.