feat(dan_layer)!: generate and add checkpoint signatures (#4261)
Description
---
- feat(dan_layer): generate and add checkpoint signatures
- feat(explorer): update text explorer to include signatures
-  feat(base_layer/wallet): update wallet command file formats
-  feat(base_layer/core): wires up committee signatures to grpc with conversions
-  feat(base_layer/core): adds SignerSignature struct for CommitteeSignatures
-  feat(base_layer/wallet): include committee signatures for checkpoint GRPC requests

Motivation and Context
---
DAN committees need to collect signatures to include in checkpoints. This PR adds the initial implementation of that.
The signature challenge is: 

`e = H_1(signer_public_key || public_nonce || H_2(contract_id||commitment||merkle_root||checkpoint_number))`

Although constructing a challenge with actual data is a TODO as that involves a bit of a refactor.

This PR also adds the `SignerSignature` struct that includes the signing public key along with the Schnorr signature tuple `<R, s>`. This is so that we know which if the validators signed. There may be an aggregate signature scheme that allows this.

How Has This Been Tested?
---
Manually, initial checkpoint contains signature.