feat(wallet): uses tip height to calc abs acceptance period (#4271)
Description
---
- fetches tip height from base node to determine the absolute height for the VN acceptance period

Motivation and Context
---
When using `init-constitution` command used an absolute height for the acceptance period, which made it easy to 
make the mistake of creating a constitution that expires immediately.  

This PR fetches the tip height from the base node and uses that, if able, to make the height relative. If not able, the user is warned and prompted to enter an ABSOLUTE height.

How Has This Been Tested?
---
Manually