feat: add quarantine and backup keys (#4289)
Description
---
This PR adds the quarantine interval and backup_keys to the constitution as per the RFC. 

This also moves the side_chain features to boxed struct.

Motivation and Context
---
Adding the quarantine interval and backup_keys makes the output_features big enough that it causes a stack overflow error in some of the wallet tests. Since we will most likely add more features to the side_chain features, this is moved to a boxed struct so that it wont fail. 

How Has This Been Tested?
---
Unit tests