test: improve test reliability (#4309)
Description
---
This test `check_share` has dynamic mining in it, which is not realiabilie testing as it can fail mining at all, or in this case, it can mine a difficulty that is higher than "fail" safe high values. 

This makes sure we change difficulties to ensure we test the intended behavior. 

How Has This Been Tested?
---
Ran the test 10_000 times without a failure.