refactor(launchpad): move launchpad to its own repo
This PR removes Tari Launchpad from the main repo. It's new home will be
tari-project/tari-launchpad.

The entire application is removed here; including the backend, front-end
and Ci workflows.

The responsibility for building the docker images associated with the
app will also reside in the tari-launchpad repo.

The reasons for the move are:
- Separation of concerns
- Speed of development. Having the full Tari test suite run for every
  launchpad PR is incredibly time-consuming and resource intensive.