feat(dan): add WASM template invocation from user instruction (#4331)
Description
---
Adds minimum required code to call a WASM template from a user instruction.

Motivation and Context
---
This PR adds and stubs out the required code to call WASM templates from a user instruction. This PR makes use of radix's excellent sbor encoding library, but this is not required and could be switched out.

It's quite messy and incomplete in places.  

How Has This Been Tested?
---
Adds an integration test that compiles and calls the stubbed-out template code.
This is bordering on POC code, so no other tests are written - tests should be implemented after some feedback/iterations from this PR.