fix: add more recent zero point for wallet birthday (see issue #4176) (#4275)
Description
--- 
Add logic to tackle issue #4176, which aims at improving birthday representation for Cipher Seed generation.

Motivation and Context
--- 
The goal of this PR is to resolve issue #4176.  The current version is supposed to allow a more recent genesis for wallet birthdays (it actually depends on the genesis block of the Tari network, we wish to work with). Moreover, it updates the version. Moreover, if the birthday does not fit within the `u16` type range, we update to a new version. 

How Has This Been Tested?
--- 
Add simple tests to check if logic is correct.