feat!: fix kernel mutability (#4377)
Description
---
Currently, the kernel has a mutability issue where the public excess, features, and new optional burn commitment is not committed to in the challenge.  
See issue: https://github.com/tari-project/tari/issues/4365

This is a breaking change as it changes every single kernel signature. 

This changes the challenge for the kernel to include those fields.  Because the kernel fields need to be signed by all parties, these needs fields must be decided at the start of the transaction. This required changes to the tx protocol as well. 

How Has This Been Tested?
---
Running all current unit tests and all critical cucumber tests.