docs: new contribution guidelines (#4163)
Provides a new set of contribution guidelines to help existing and new
contributors understand the process and help them in getting code merged
into the code base more quickly.

The focus of this rewrite is:
* to introduce some recommendations for a post-mainnet Tari
* Outline the release process post-mainnet
* Outline the PR review and evaluation process post-mainnet (but
  starting to implement immediately)
* to provide guidelines that result in contributions that are easy to
  review, allowing reviewers to catch more bugs.

## Important 

This PR is a starting point for a discussion around these guidelines. This is not me coming down from the mountain with a set of clay tablets. I've done a fair bit of research and thinking, and also consulted with many other devs, including @fluffypony in reaching this draft. However I expect that the final version, when it is merged, will be somewhat different from what is written here now, but my hope is that the spirit of it remains.