chore: remove wallet key manager state migration (#4384)
Description
---
Removes the wallet key manager state migration.
Removes key_manager_state_old db.

Motivation and Context
---
During the dibbler testnet, we changed how the key manager state is managed and used. This required us to recreate how the wallet stored the key manager states in the database. This required us to keep the old state and migrate over to the new storage scheme for the older wallets already running on the old one. Because we are resetting the testnet, we won't have older wallets, so we can remove the migration code.

Fixes: https://github.com/tari-project/tari/issues/4354