fix: fix new block handling of non-tip blocks (#4431)
Description
---
Currently, the base node cannot process a block that was received out of order (aka orphaned).
The `reconcile_block` function makes the assumption that the block is always on the current tip and tries to calculate the MMR roots of the block as a sanity check. The MMR roots can only be checked by a block that is built on the current local tip. 
This changes the `reconcile_block` function to first check orphan status, and if true, it will require the full block from the peer. 

This Pr also changes how the db lock is processed. Currently, it will print out a warn followed by an error. This suppresses the error and only prints the warn. 

How Has This Been Tested?
---
Manual


Fixes: https://github.com/tari-project/tari/issues/4352