fix(base_node_config): check_interval is 0 made base node is panicked (#4495)
Description
--
This PR served for fixing the bug described in #4399 

Motivation and Context
---
My target is finding a way to change it with as little change and minimal impact as possible. So here's how I can think of it. I'm not sure if this is the best solution. Any comments are welcome.

How Has This Been Tested?
---
Build runs in local fork, with `check_interval` = 0 and w/o `check_interval` setting.