fix: grpc inconsistent serialization of keys (see issue #4224) (#4491)
Description
--- 
We address inconsistent serialization of keys across base node and wallet grpc servers. The first uses a little endian 32-byte representation of the keys, whereas the second a 64-byte hex representation. We opt to use a hex-representation for the public key on the wallet grpc side.

Motivation and Context
--- 
Fixes #4224.

How Has This Been Tested?
--- 
Manually.