fix(comms): ignore dial cancel event when inbound connection exists (#4521)
Description
---
Ignores `DialCancelled` event when the dialer is notified about an established `Inbound` connection

Motivation and Context
---
In the case where A dials B and B dials A and the B->A(Inbound relative to A) connection occurs first. The connectivity manager notifies the dialer actor to cancel any pending dials. However, this emits a DialCancelled event which causes the connectivity manager to discard the previous inbound connection.

This PR ignores canceled dial events when an active inbound connection exists.

Difficult to outright confirm, but I suspect this fixes:
- dial/disconnect churn
- A and B disagreeing on being connected to each other for a time/indefinitely.

How Has This Been Tested?
---
Hard to test for, tested manually by attempting to make B -> A happen first and checking that this case occurs in the logs.