feat!: change how sha3 difficulty is calculated (#4528)
Description
---
Change how the sha3 hash for the difficulty is calculated. 
Before this pr the difficulty was calculated as the Hash(header.part || header.part || ... || header.part), which is then hashed again. 

The merge mining is calculated as the hash of all the parts bar the nonce and pow data. This hash is then concatenated with the nonce and pow data. This makes it much easier to send over data to hash as only this hash of data is required to mine the header and not the entire header. 

This updates the sha3 mining to also use this same hash rather than the concatenated header part. 

How Has This Been Tested?
---
unit tests