feat: show the banned reason if a peer was banned in the contacts (#4525)
Description
---
Currently, if a peer has been banned, the constants will only show the peer is offline. 
This PR queries the peer manager for the peer to ask if the peer was banned or not. When asking the contact service in the wallet for the online status of the contact, the service will now report if the peer was banned, as well the reason for the banning.