fix(core)!: replace block and transaction hashes with FixedHash type (#4533)
Description
---
This removes the usage of Vec<u8> for all hashes except the mmr and RPC requests. 
All other internal hashing now uses the `FixedHash` type. 
We now know each hash will always be of the correct length when processing the hash.
This gets rid of a good amount of technical debt. 

How Has This Been Tested?
---
unit tests