Improve recovery of coinbase (#4604)
Description
---
This PR improves how the wallet displays and handles recovered coinbases. 

Motivation and Context
---
Most of the coinbase transactions, have all of the required information to properly import then with the correct info on the utxo, so we don't have to flag them as imported. 

How Has This Been Tested?
---
Manual 

Fixes: https://github.com/tari-project/tari/issues/4581