feat: make sure duplication check happens first in mempool (#4627)
Description
---
Changes the order of validation in mempool to check the excess signature before checking the inputs. 

Motivation and Context
---
When receiving old transactions, if we check the inputs first, the tx will fail on input validation, and we flag this as a transaction that double spends some input. 
This is incorrect as it's a rebroadcast of an old already mined transaction. So we check the kernel_excess signature first, and if this is already contained in the blockchain, we just print out a debug message saying we received an old already mined kernel.