fix: fee estimate (#4656)
Description
---
Allows fee estimate to always return an estimated fee. 

Motivation and Context
---
Currently, when calling fee estimate and you do not have enough funds, or funds pending, the fee estimate will error. 
We do not require such a high level of accuracy when calculating a fee estimate for an external call for an estimation. 
When the wallet does not have enough funds available, the API will return the fee estimate for default with 1 input and 1 kernel. 


How Has This Been Tested?
---
unit test