fix(comms): simplify and remove possibility of deadlock from pipelines and substream close (#4676)
Description
---
- Simplify outbound pipeline by removing the [pipeline] -> [messaging] channel
- Pipe outbound messages directly to the messaging protocol instead of through the outbound pipeline
- Fix rare lockup when calling yamux control close 

Motivation and Context
---
The outbound pipeline needed to poll two channels in order to make progress, some code branches in the outbound pipeline may need to use the other channel, and if that channel is full and the number of concurrent outbound tasks are full, a deadlock will occur. This case has not been directly observed, but is technically possible so should be eliminated.

This PR removes the [pipeline] -> [messaging] channel, making the outbound pipeline only have to poll one channel. It also directly pipes `OutboundMessage`s to the messaging protocol.

EDIT: I believe I've found the root cause. The connectivity manager would rarely "lock up" causing the pipelines to lock up (both pipelines require calls to connectivity manager). I traced this in the logs and found that the last thing the connectivity manager does is resolve a tie break before locking up. This involves disconnecting one of the peer connections, and it appeared this future, extremely rarely, did not resolve. Digging deeper from there, I was able to track down a flaw in the substream close procedure, write a test that reproduces it and make a fix.

How Has This Been Tested?
---
Number of ~1000-2000tx stress tests, leaving base nodes overnight (none of these are conclusive but no issues were encountered)