fix: sync handling and increase reorg speed in mempool (#4706)
Description
---
This PR updates the handling reorgs and syncs in the mempool. 

The reorg code is updated to remove unnecessary checks when handling reorgs.

The sync code is completely redone to now only listen for a sync completed status. As of now the mempool ignores the rewind event as the mempool knows this is not the final state, either the base node will forward to a new tip or reset itself to the old tip. The base node upon receiving sync completed event will re-evaluate all transactions inside of it and reset it state to the correct state. 

There will be a followup pr to add mempool sync