feat: improve bn command mode timeouts (#4712)
Description
---
Assigned a base node command mode time-out based on the typical time it will take to return back to the command line for specific commands.

Motivation and Context
---
Differentiation in the time-out is needed for the various commands, for example, `check-db` and `rewind-blockchain` need a lot more time to complete than `version` or `whoami`. The specific case encountered was that the previous blanket allowed time-out of `70s` was not sufficient to perform `check-db`  or `get-mempool-stats` for a busy (stressed) block chain and mempool.

How Has This Been Tested?
---
System-level testing.