fix: cli wallet cucumber (#4739)
Description
---
Fix the burn cli cucumber.
The command run for wallet didn't start with grpc. And the wallet in non-interactive mode without grpc will not start. So it was not in sync with the base node.
Modify the standard cucumber burn test to do the same as the cli test.

How Has This Been Tested?
---
npm test -- --name "As a user I want to burn tari via command line"