feat: adds a create combo key cli command for wallet (#4751)
Description
---
Adds a `create_combo_key` method to wallet. The main goal is to allow the user to create a new pair of (secret key, public key), by passing in a given string value.

Motivation and Context
---
Wallet needs a terminal command to create a new public key and secret key from the key manager service.
This key needs to be created based on a provided string seed and send back to the user. 

The key manager service derives a key based on the string and the master seed of the wallet. We need to add this seed string as a branch on the local instance of the key manager and then we can track if the user asks for the key twice.

The current command is needed in order to as a firs step to execute an m-of-n multi-party script payment protocol. See [here](https://demo.hedgedoc.org/rgLUGP_YRiGKjKkgehiG4w#Step-2) and [here](https://demo.hedgedoc.org/0Lb9sKovSb-ct3g3E4A-9A?view) for more details.

How Has This Been Tested?
---
Unit tests